ref(credential request) making file location amendments
