+++ frontend/src/components/InteractionBtn/InteractionBtn.module.css
@@ -15,4 +15,5 @@
   width: 1.25rem;
   margin-right: 0.5rem;
   flex: 1;
+  width: '100%';
 }
