+++ frontend/src/components/InteractionCredentialSelect/InteractionCredentialSelect.module.css
@@ -1,0 +1,9 @@
+.container {
+  padding-top: '20px';
+}
+
+.item {
+  border-radius: 10px;
+  margin: 10px;
+  padding: 10px;
+}
