+++ frontend/src/components/InteractionInput/Input.module.css
@@ -1,7 +1,3 @@
-.input-container {
-  padding-top: 20px;
-}
-
 .input {
   margin: 10px;
   width: 100%;
