+++ frontend/src/components/Space.tsx
@@ -1,0 +1,3 @@
+import React from 'react'
+
+export const Space = () => <div style={{ paddingTop: '20px' }} />
