+++ frontend/src/ui/interaction.tsx
@@ -101,59 +101,3 @@ export const CredOfferContainer = ({
     </InteractionTemplate>
   )
 }
-
-export const CredShareContainer = ({
-  serviceAPI,
-  credTypes,
-}: {
-  serviceAPI: JolocomWebServiceClient
-  credTypes: string[]
-}) => {
-  const [requestedCredentials, setRequested] = useState<Array<string>>([])
-  const requestableCredTypes = credTypes
-
-  const handleSelect = (array: string[], item: string) => {
-    return !array.includes(item)
-      ? [...array, item]
-      : array.filter(val => val !== item)
-  }
-
-  useEffect(() => {
-    if (
-      requestedCredentials.length === 0 &&
-      credTypes &&
-      credTypes.length > 0
-    ) {
-      setRequested(credTypes.slice(0, 1))
-    }
-  })
-
-  const startCredRequest = async () => {
-    const resp: { qr: string; err: string } = await serviceAPI.sendRPC(
-      RpcRoutes.credShareRequest,
-      {
-        types: Array.from(new Set(requestedCredentials)),
-      },
-    )
-    console.log(resp)
-    return resp
-  }
-
-  return (
-    <InteractionTemplate
-      startText="Start Credential Request Interaction"
-      startHandler={startCredRequest}
-    >
-      <h2>Credential Request</h2>
-      <SelectionComponent
-        title={'Available Credentials'}
-        options={requestableCredTypes}
-        onSelect={type =>
-          setRequested(handleSelect(requestedCredentials, type))
-        }
-        selectedItems={requestedCredentials}
-      />
-      <div style={{ paddingTop: '20px' }}></div>
-    </InteractionTemplate>
-  )
-}
